Make CLI backward compatible
